Adding the Update Rest Api for Shopping Cart

+++ src/main/java/com/example/ecommerceapp/repo/ProductRepo.java
@@ -5,7 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepo extends JpaRepository<Product, Long>{
+public interface ProductRepo extends JpaRepository<Product, Integer>{
 
-   // Product updateProduct();
+
+   // void deleteProductById(Integer id);
+
+
+    // Product updateProduct();
 }
